refactor(WeekDay,WeekDayTest): WeekDay 클래스 이름 변경
* 평일과 이름이 겹쳐 요일을 의미하도록 WeekDay -> DayOfWeek 으로 이름 변경
