+++ src/test/java/christmas/DayOfWeekTest.java
@@ -1,21 +1,21 @@
 package christmas;
 
-import christmas.domain.WeekDay;
+import christmas.domain.DayOfWeek;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class WeekDayTest {
+public class DayOfWeekTest {
     @ParameterizedTest
     @ValueSource(ints = {1, 9})
     void 주말_요일_테스트(int input) {
-        assertThat(WeekDay.isHoliday(input)).isTrue();
+        assertThat(DayOfWeek.isHoliday(input)).isTrue();
     }
 
     @ParameterizedTest
     @ValueSource(ints = {3, 13})
     void 평일_요일_테스트(int input) {
-        assertThat(WeekDay.isHoliday(input)).isFalse();
+        assertThat(DayOfWeek.isHoliday(input)).isFalse();
     }
 }
